requirements.txt
streamlit
streamlit-feedback
trubrics
openai
requests
python-dotenv
pydantic==1.*
pandas
certifi
langchain
langchain_openai
bs4
tiktoken